store/tikv: resolve util dependencies
