+++ util/misc.go
@@ -47,8 +47,6 @@ const (
 	DefaultMaxRetries = 30
 	// RetryInterval indicates retry interval.
 	RetryInterval uint64 = 500
-	// GCTimeFormat is the format that gc_worker used to store times.
-	GCTimeFormat = "20060102-15:04:05 -0700"
 )
 
 // RunWithRetry will run the f with backoff and retry.
@@ -123,26 +121,6 @@ func Recover(metricsLabel, funcInfo string, recoverFn func(), quit bool) {
 		time.Sleep(time.Second * 15)
 		os.Exit(1)
 	}
-}
-
-// CompatibleParseGCTime parses a string with `GCTimeFormat` and returns a time.Time. If `value` can't be parsed as that
-// format, truncate to last space and try again. This function is only useful when loading times that saved by
-// gc_worker. We have changed the format that gc_worker saves time (removed the last field), but when loading times it
-// should be compatible with the old format.
-func CompatibleParseGCTime(value string) (time.Time, error) {
-	t, err := time.Parse(GCTimeFormat, value)
-
-	if err != nil {
-		// Remove the last field that separated by space
-		parts := strings.Split(value, " ")
-		prefix := strings.Join(parts[:len(parts)-1], " ")
-		t, err = time.Parse(GCTimeFormat, prefix)
-	}
-
-	if err != nil {
-		err = errors.Errorf("string \"%v\" doesn't has a prefix that matches format \"%v\"", value, GCTimeFormat)
-	}
-	return t, err
 }
 
 // HasCancelled checks whether context has be cancelled.
